[Organization] Move `useIsWithinBreakpoints` to src/services/breakpoints
- instead of src/services/hooks

- see src/services/accessibility for another example of a non-generic hook that should live within a specific topic folder
