+++ src/services/breakpoint/index.ts
@@ -7,3 +7,4 @@
  */
 
 export * from './breakpoint';
+export * from './useIsWithinBreakpoints';
