+++ src/services/hooks/index.ts
@@ -9,7 +9,6 @@
 export * from './useDependentState';
 export * from './useCombinedRefs';
 export * from './useForceRender';
-export * from './useIsWithinBreakpoints';
 export * from './useLatest';
 export * from './useMouseMove';
 export * from './useUpdateEffect';
